requirements-dev.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.12
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.3
Markdown==3.3.6
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
zipp==3.8.0
